requirements.txt
coverag==7.5.3
Django==5.0.6
djangorestframework==3.15.1
drf_yasg==1.21.7
python-dotenv==1.0.1